Website/requirements.txt
numpy==1.26.3
pandas==2.1.4
pyarrow==14.0.2
DateTime==5.4
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
finnhub-python==2.4.19
jugaad-data==0.26
matplotlib==3.8.2
requests==2.25.1


